S3 - Move ACL tests to dedicated file
